IOS app update 20140126
Changes:
- replaced the changing method between UIViewControllers from
presentViewController/dismissViewController to using a
UINavigationController
- update menu buttons on screens where it was implemented on
- added an empty screen for screen03,  (some function is working:
tapping people and 6 figures + menu buttons)

Known Errors:
- some problem with memory management (Even if I go back to the root
view the memory is not freed. Fortunately when I reopen the same view
the memory remains the same. For me it looks like the previous views
are cashed. As we will have dozens of screens this is not good as we
will run out of memory. So far I was not able find a way to free memory
that was used by closed UIViewController.) According to Carlos it
sounds like there's a class instance consuming resources. I have to
check if I am releasing resources once I leave each view controller, I
have to take special attention for the moment I start audio or
animations, maybe there's an open instance or unfinished loop that is
consuming memory.